+++ src/driver.rs
@@ -3,7 +3,6 @@ use std::path::Path;
 use std::sync::Once;
 
 use gdal_sys::{self, CPLErr, GDALDriverH, GDALMajorObjectH};
-use libc::c_int;
 
 use crate::cpl::CslStringList;
 use crate::dataset::Dataset;
@@ -105,9 +104,9 @@ impl Driver {
     pub fn create<P: AsRef<Path>>(
         &self,
         filename: P,
-        size_x: isize,
-        size_y: isize,
-        bands: isize,
+        size_x: usize,
+        size_y: usize,
+        bands: usize,
     ) -> Result<Dataset> {
         self.create_with_band_type::<u8, _>(filename, size_x, size_y, bands)
     }
@@ -134,9 +133,9 @@ impl Driver {
     pub fn create_with_band_type<T: GdalType, P: AsRef<Path>>(
         &self,
         filename: P,
-        size_x: isize,
-        size_y: isize,
-        bands: isize,
+        size_x: usize,
+        size_y: usize,
+        bands: usize,
     ) -> Result<Dataset> {
         let options = [];
         self.create_with_band_type_with_options::<T, _>(filename, size_x, size_y, bands, &options)
@@ -176,9 +175,9 @@ impl Driver {
     pub fn create_with_band_type_with_options<T: GdalType, P: AsRef<Path>>(
         &self,
         filename: P,
-        size_x: isize,
-        size_y: isize,
-        bands: isize,
+        size_x: usize,
+        size_y: usize,
+        bands: usize,
         options: &[RasterCreationOption],
     ) -> Result<Dataset> {
         Self::_create_with_band_type_with_options(
@@ -195,9 +194,9 @@ impl Driver {
     fn _create_with_band_type_with_options(
         &self,
         filename: &Path,
-        size_x: isize,
-        size_y: isize,
-        bands: isize,
+        size_x: usize,
+        size_y: usize,
+        bands: usize,
         data_type: GdalDataType,
         options: &[RasterCreationOption],
     ) -> Result<Dataset> {
@@ -206,14 +205,18 @@ impl Driver {
             options_c.set_name_value(option.key, option.value)?;
         }
 
+        let size_x = libc::c_int::try_from(size_x)?;
+        let size_y = libc::c_int::try_from(size_y)?;
+        let bands = libc::c_int::try_from(bands)?;
+
         let c_filename = _path_to_c_string(filename)?;
         let c_dataset = unsafe {
             gdal_sys::GDALCreate(
                 self.c_driver,
                 c_filename.as_ptr(),
-                size_x as c_int,
-                size_y as c_int,
-                bands as c_int,
+                size_x,
+                size_y,
+                bands,
                 data_type as u32,
                 options_c.as_ptr(),
             )