+++ CHANGES.md
@@ -11,6 +11,15 @@
 
    - <https://github.com/georust/gdal/pull/494>
 
+- **Breaking**: Changed a number of APIs using `isize` when `usize` is semantically more appropriate: `Driver::create.*`, `Rasterband::overview`, `Dataset::{layer|into_layer|layer_count}`.  
+
+  - <https://github.com/georust/gdal/pull/497>
+
+- Created `enum AxisMappingStrategy` for `OSRAxisMappingStrategy` ordinals.
+- **Breaking**: `SpatialRef::{set_}axis_mapping_strategy` use `AxisMappingStrategy` instead of `gdal_sys::OSRAxisMappingStrategy::Type`.
+
+   - <https://github.com/georust/gdal/pull/498>
+
 - Defers the `gdal_i.lib` missing message until after the `pkg-config` check and outputs `pkg-config` metadata in case of a static build.
 
    - <https://github.com/georust/gdal/pull/492>